add MessageStatus component
